Introduced the F amode for sign-extended immediates
This does not affect any existing behavior.  The immediate operands now
feature an "extended" field that contains the sign-extended immediate
value.

There are only a few instructions that use this, notably PUSH imm8,
three-operand IMUL and the M1 ModR/M map.

Printing has not been altered and will output the way the immediate was
encoded.